requirements.txt
openai>=1.54.0
langchain>=0.3.7
langchain-openai>=0.2.9
python-dotenv>=0.21.0